+++ harness/evals/integration/schemas/authored-scenario.v1.json
