+++ harness/evals/integration/schemas/compiled-scenario.v1.json
@@ -1,0 +1,285 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "additionalProperties": false,
+  "definitions": {
+    "CompiledFaultV1": {
+      "additionalProperties": false,
+      "properties": {
+        "after_target_calls": {
+          "format": "uint64",
+          "minimum": 1.0,
+          "type": "integer"
+        },
+        "function_id": {
+          "type": "string"
+        },
+        "kind": {
+          "$ref": "#/definitions/FaultKind"
+        },
+        "restart_delay_ms": {
+          "format": "uint64",
+          "minimum": 0.0,
+          "type": "integer"
+        }
+      },
+      "required": [
+        "after_target_calls",
+        "function_id",
+        "kind",
+        "restart_delay_ms"
+      ],
+      "type": "object"
+    },
+    "DeadlinesV1": {
+      "additionalProperties": false,
+      "properties": {
+        "readiness_ms": {
+          "default": 60000,
+          "format": "uint64",
+          "minimum": 1.0,
+          "type": "integer"
+        },
+        "scenario_ms": {
+          "default": 60000,
+          "format": "uint64",
+          "minimum": 1.0,
+          "type": "integer"
+        },
+        "teardown_ms": {
+          "default": 15000,
+          "format": "uint64",
+          "minimum": 1.0,
+          "type": "integer"
+        }
+      },
+      "type": "object"
+    },
+    "FaultKind": {
+      "enum": [
+        "engine_sigkill"
+      ],
+      "type": "string"
+    },
+    "InvariantSpecV1": {
+      "additionalProperties": false,
+      "description": "Stable wire form persisted in compiled scenarios.",
+      "properties": {
+        "id": {
+          "type": "string"
+        },
+        "parameters": {
+          "additionalProperties": true,
+          "default": {},
+          "type": "object"
+        }
+      },
+      "required": [
+        "id"
+      ],
+      "type": "object"
+    },
+    "LifecycleFunctionId": {
+      "enum": [
+        "integration-recorder::lifecycle"
+      ],
+      "type": "string"
+    },
+    "LifecycleTriggerType": {
+      "enum": [
+        "harness::turn-started",
+        "harness::turn-completed"
+      ],
+      "type": "string"
+    },
+    "RecorderConfigV1": {
+      "additionalProperties": false,
+      "properties": {
+        "extra_functions": {
+          "description": "Additional run-scoped controlled functions (e.g. hook implementations for `harness::hook::*` scenarios). Registered exactly like the target: declared description/schema verbatim, every call durably recorded, declared response returned.",
+          "items": {
+            "$ref": "#/definitions/RecorderTargetV1"
+          },
+          "type": "array"
+        },
+        "lifecycle": {
+          "$ref": "#/definitions/RecorderLifecycleV1"
+        },
+        "target": {
+          "$ref": "#/definitions/RecorderTargetV1"
+        }
+      },
+      "required": [
+        "lifecycle",
+        "target"
+      ],
+      "type": "object"
+    },
+    "RecorderLifecycleV1": {
+      "additionalProperties": false,
+      "properties": {
+        "function_id": {
+          "$ref": "#/definitions/LifecycleFunctionId"
+        },
+        "trigger_type": {
+          "$ref": "#/definitions/LifecycleTriggerType"
+        }
+      },
+      "required": [
+        "function_id",
+        "trigger_type"
+      ],
+      "type": "object"
+    },
+    "RecorderTargetV1": {
+      "additionalProperties": false,
+      "properties": {
+        "description": {
+          "type": "string"
+        },
+        "function_id": {
+          "description": "Must be prefixed by `<run_id>::` — enforced by `configure`.",
+          "type": "string"
+        },
+        "request_schema": {
+          "additionalProperties": true,
+          "description": "Registered verbatim; the same schema must appear in native tool exposure (that registration is part of the oracle).",
+          "type": "object"
+        },
+        "response": {
+          "description": "Declared response returned for every target call."
+        },
+        "response_delay_ms": {
+          "description": "Delay between the durable append and the response, opening a window for fault injection while the dispatched call is executing (crash-recovery scenarios). Omitted = respond immediately.",
+          "format": "uint64",
+          "minimum": 0.0,
+          "type": [
+            "integer",
+            "null"
+          ]
+        }
+      },
+      "required": [
+        "description",
+        "function_id",
+        "request_schema",
+        "response"
+      ],
+      "type": "object"
+    },
+    "ReleaseActionV1": {
+      "enum": [
+        "execute",
+        "deliver"
+      ],
+      "type": "string"
+    },
+    "ReleaseV1": {
+      "additionalProperties": false,
+      "properties": {
+        "action": {
+          "$ref": "#/definitions/ReleaseActionV1"
+        },
+        "function_call_id": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "action"
+      ],
+      "type": "object"
+    },
+    "SchemaVersion1": {
+      "description": "The literal string `\"1\"`; any other value is a schema error.",
+      "enum": [
+        "1"
+      ],
+      "type": "string"
+    },
+    "TriggerBindingV1": {
+      "additionalProperties": false,
+      "properties": {
+        "config": true,
+        "function_id": {
+          "type": "string"
+        },
+        "trigger_type": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "config",
+        "function_id",
+        "trigger_type"
+      ],
+      "type": "object"
+    }
+  },
+  "description": "Strict runtime scenario produced from [`super::IntegrationScenarioV1`].",
+  "properties": {
+    "bindings": {
+      "items": {
+        "$ref": "#/definitions/TriggerBindingV1"
+      },
+      "type": "array"
+    },
+    "deadlines": {
+      "$ref": "#/definitions/DeadlinesV1"
+    },
+    "description": {
+      "type": "string"
+    },
+    "fault": {
+      "anyOf": [
+        {
+          "$ref": "#/definitions/CompiledFaultV1"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
+    "id": {
+      "maxLength": 128,
+      "minLength": 1,
+      "pattern": "^[A-Za-z0-9_-]+$",
+      "type": "string"
+    },
+    "invariants": {
+      "items": {
+        "$ref": "#/definitions/InvariantSpecV1"
+      },
+      "type": "array"
+    },
+    "quarantine": {
+      "type": "boolean"
+    },
+    "recorder": {
+      "$ref": "#/definitions/RecorderConfigV1"
+    },
+    "release": {
+      "anyOf": [
+        {
+          "$ref": "#/definitions/ReleaseV1"
+        },
+        {
+          "type": "null"
+        }
+      ]
+    },
+    "schema_version": {
+      "$ref": "#/definitions/SchemaVersion1"
+    },
+    "send": true
+  },
+  "required": [
+    "deadlines",
+    "description",
+    "id",
+    "invariants",
+    "recorder",
+    "schema_version",
+    "send"
+  ],
+  "title": "CompiledScenarioV1",
+  "type": "object"
+}
